🧬 Extract the details logic of subcommand line *check* into another single module.
